Move ChannelsIndex into channels folder
