+++ web/static/js/components/channels/ChannelIndex.jsx
@@ -2,14 +2,14 @@
 import React, { Component, PropTypes } from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { orderedItems } from '../reducers/collection'
-import * as actions from '../actions/channels'
+import { orderedItems } from '../../reducers/collection'
+import * as actions from '../../actions/channels'
 import range from 'lodash/range'
-import * as authActions from '../actions/authorizations'
-import { AddButton, EmptyPage, CardTable, UntitledIfEmpty, SortableHeader, Modal, ConfirmationModal } from './ui'
+import * as authActions from '../../actions/authorizations'
+import { AddButton, EmptyPage, CardTable, UntitledIfEmpty, SortableHeader, Modal, ConfirmationModal } from '../ui'
 import { Preloader } from 'react-materialize'
-import { config } from '../config'
-import { channelFriendlyName } from '../channelFriendlyName'
+import { config } from '../../config'
+import { channelFriendlyName } from '../../channelFriendlyName'
 import { translate } from 'react-i18next'
 
 class ChannelIndex extends Component {
