+++ web/static/js/routes.jsx
@@ -9,7 +9,7 @@ import SurveySettings from './components/surveys/SurveySettings'
 import SurveySimulation from './components/surveys/SurveySimulation'
 import QuestionnaireIndex from './components/questionnaires/QuestionnaireIndex'
 import QuestionnaireEditor from './components/questionnaires/QuestionnaireEditor'
-import ChannelIndex from './components/ChannelIndex'
+import ChannelIndex from './components/channels/ChannelIndex'
 import ProjectTabs from './components/projects/ProjectTabs'
 import SurveyTabs from './components/surveys/SurveyTabs'
 import RespondentIndex from './components/respondents/RespondentIndex'
